refactor(schedule): move the stop lifetime into the runner
Keep store setup on every held-lock path through the CLI callback.
